+++ src/subaction_animations.jl
@@ -36,6 +36,11 @@ function _appear(video, action, subaction, rel_frame, symbol::Val{:fade})
     action.current_setting.mul_opacity = t
 end
 
+function _appear(video, action, subaction, rel_frame, symbol::Val{:scale})
+    t = get_interpolation(subaction, rel_frame)
+    action.current_setting.mul_scale = t
+end
+
 """
     disappear(s::Symbol)
 
@@ -72,6 +77,11 @@ end
 function _disappear(video, action, subaction, rel_frame, symbol::Val{:fade})
     t = get_interpolation(subaction, rel_frame)
     action.current_setting.mul_opacity = 1 - t
+end
+
+function _disappear(video, action, subaction, rel_frame, symbol::Val{:scale})
+    t = get_interpolation(subaction, rel_frame)
+    action.current_setting.mul_scale = 1 - t
 end
 
 """