+++ test/unit.jl
@@ -189,6 +189,89 @@
         @test action.internal_transitions[1].center == Point(2.0, 5.0)
     end
 
+    @testset "scaling" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(0:100, () -> 1, Scaling(0.0, 1.0))
+        # needs internal scaling as well
+        push!(action.internal_transitions, Javis.InternalScaling((0, 0)))
+        Javis.compute_transformation!(action, video, 0)
+        @test action.internal_transitions[1].scale == (0.0, 0.0)
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].scale == (0.5, 0.5)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].scale == (1.0, 1.0)
+    end
+
+    @testset "Relative frames" begin
+        video = Video(500, 500)
+        action = Action(Rel(10), (args...) -> 1, Translation(Point(1, 1), Point(100, 100)))
+        # dummy action doesn't need a real function
+        test_file = javis(
+            video,
+            [
+                Action(1:100, (args...) -> 1, Translation(Point(1, 1), Point(100, 100))),
+                action,
+            ],
+        )
+        @test Javis.get_frames(action) == 101:110
+        rm(test_file)
+    end
+
+    @testset "translation from origin" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Translation(Point(99, 99)))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == O
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(49, 49)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(99, 99)
+
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Translation(99, 99))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == O
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(49, 49)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(99, 99)
+    end
+
+    @testset "rotations" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Rotation(2π))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalRotation(0.0, O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].angle == 0.0
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].angle == 2π
+
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Rotation(2π, Point(2.0, 5.0)))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalRotation(0.0, O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].angle == 0.0
+        @test action.internal_transitions[1].center == Point(2.0, 5.0)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].angle == 2π
+        @test action.internal_transitions[1].center == Point(2.0, 5.0)
+    end
+
     @testset "Frames errors" begin
         video = Video(500, 500)
         # throws because the frames of the first action need to be defined explicitly