+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Ability to use [Animations.jl](https://github.com/jkrumbiegel/Animations.jl) 
   - for Transformations and `appear` and `disappear`
   
+## Unreleased
+- Ability to scale an object with `Scaling`. Works similar to `Translation` and `Rotation` 
+- Added JuliaFormatter GitHub Action
 
 ## 0.1.2 (24th of August 2020)
 - Added capabilities for generating `.mp4` files